+++ src/Components/auth/AuthDetails/AuthDetails.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { auth } from "./firebase";
+import { auth } from "../firebase";
 import { onAuthStateChanged, signOut } from "firebase/auth";
 import { useNavigate } from "react-router-dom";
