+++ src/Components/auth/Signin/Signin.scss
@@ -1,0 +1,69 @@
+@use "../../../styles/partials/mixins" as *;
+@use "../../../styles/partials/variables" as *;
+
+.signIn {
+  background-color: $neutral;
+  padding: 2rem;
+  border-radius: 10px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+
+  margin: 0 auto;
+  text-align: center;
+
+  &__loginHeader {
+    color: $charcoal;
+    font-size: 1.5rem;
+    padding-bottom: 1rem;
+
+    @include tablet {
+      font-size: 2rem;
+    }
+    @include desktop {
+      font-size: 3rem;
+    }
+  }
+
+  &__title {
+    color: $charcoal;
+    font-size: 2rem;
+    margin-bottom: 1.5rem;
+  }
+
+  &__email,
+  &__password {
+    width: 100%;
+    padding: 0.75rem;
+    margin-bottom: 1.25rem;
+    border: 1px solid $charcoal;
+    border-radius: 5px;
+    font-size: 1rem;
+    background-color: $paleBlue;
+
+    &::placeholder {
+      color: $charcoal;
+      opacity: 0.7;
+    }
+
+    &:focus {
+      border-color: $indigoBlue;
+      outline: none;
+      box-shadow: 0 0 5px rgba($indigoBlue, 0.5);
+    }
+  }
+
+  &__login {
+    background-color: $mint;
+    color: $charcoal;
+    font-size: 1.1rem;
+    padding: 0.75rem;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+
+    &:hover {
+      background-color: $coral;
+      color: $neutral;
+    }
+  }
+}
